Split out `srl backup` (#112)
* split out backup to multiple files

* edit wording
