+++ srl/commands/nextup/list_.py
@@ -1,7 +1,7 @@
 from rich.console import Console
 from rich.panel import Panel
 from srl.utils import format_problem
-from .util import get_next_up_problems
+from .utils import get_next_up_problems
 
 
 def add_subparser(subparsers):
